--- pruebas-cypress/cypress.config.js
@@ -1,15 +1,0 @@
-const { defineConfig } = require("cypress");
-
-module.exports = defineConfig({
-  e2e: {
-    baseUrl: 'http://localhost:2368/"',
-    specPattern: 'cypress/e2e/**/*.cy.{js,jsx,ts,tsx}',
-
-    setupNodeEvents(on, config) {
-      // implement node event listeners here
-    },
-    env: {
-      "baseUrl" : "http://localhost:2368/",
-    }    
-  },
-});
