--- pruebas-cypress/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "cypress-boilerplate",
-  "version": "1.1.0",
-  "description": "Boilerplate for cypress project",
-  "author": "Grupo Pruebas MISO",
-  "sideEffects" : false,
-  "scripts": {
-    "cy:open": "cypress open",
-    "cy:run": "cypress run"
-  },  
-  "devDependencies": {
-    "cypress": "^13.5.0"
-  }
-}
